@@ -53,10 +53,9 @@ class GarbageCollection extends CronJob {
 	/**
 	 * Disabled because apache has isolated tmp folder
 	 *
-	 * @throws \go\core\exception\ConfigurationException
 	 */
 	private function tmpFiles() {
-		$garbage =go()->getTmpFolder()->find(
+		$garbage = go()->getTmpFolder()->find(
 			[
 				'older' => new DateTime("-1 day"),
 				'empty' => true
