@@ -3,8 +3,10 @@ namespace go\core\db;
 
 use go\core\data\ArrayableInterface;
 use go\core\ErrorHandler;
-use PDO;
 use Exception;
+use JsonSerializable;
+use PDOException;
+use PDOStatement;
 
 /**
  * PDO Statement
@@ -12,7 +14,7 @@ use Exception;
  * Represents a prepared statement and, after the statement is executed, an
  * associated result set.
  */
-class Statement extends \PDOStatement implements \JsonSerializable, ArrayableInterface{
+class Statement extends PDOStatement implements JsonSerializable, ArrayableInterface{
 	
 	private $query;
 	
@@ -20,7 +22,7 @@ class Statement extends \PDOStatement implements \JsonSerializable, ArrayableInt
 		return $this->fetchAll();
 	}
 
-	public function toArray($properties = null)
+	public function toArray(array $properties = null): array
 	{
 		return $this->fetchAll();
 	}
@@ -41,7 +43,8 @@ class Statement extends \PDOStatement implements \JsonSerializable, ArrayableInt
 	 *  
 	 * @return Query
 	 */
-	public function getQuery() {
+	public function getQuery(): Query
+	{
 		return $this->query;
 	}
 	
@@ -69,64 +72,44 @@ class Statement extends \PDOStatement implements \JsonSerializable, ArrayableInt
 		}
 	}
 
-  /**
-   * Executes a prepared statement
-   *
-   * @param $input_parameters An array of values with as many elements as there are bound parameters in the SQL statement being executed. All values are treated as PDO::PARAM_STR.
-   *
-   * Multiple values cannot be bound to a single parameter; for example, it is not allowed to bind two values to a single named parameter in an IN() clause.
-   *
-   * Binding more values than specified is not possible; if more keys exist in input_parameters than in the SQL specified in the PDO::prepare(), then the statement will fail and an error is emitted.
-   *
-   * @return bool
-   * @throws Exception
-   */
-	public function execute($input_parameters = null)
+	/**
+	 * Executes a prepared statement
+	 *
+	 * @param array|null $params An array of values with as many elements as there are bound parameters in the SQL statement being executed. All values are treated as PDO::PARAM_STR.
+	 *
+	 * Multiple values cannot be bound to a single parameter; for example, it is not allowed to bind two values to a single named parameter in an IN() clause.
+	 *
+	 * Binding more values than specified is not possible; if more keys exist in input_parameters than in the SQL specified in the PDO::prepare(), then the statement will fail and an error is emitted.
+	 *
+	 * @return bool
+	 */
+	public function execute($params = null): bool
 	{
 		try {
 
-			if(isset($input_parameters) && isset($this->build['params'])) {
+			if(isset($params) && isset($this->build['params'])) {
 				$keys = array_keys($this->build['params']);
-				foreach($input_parameters as $v) {
+				foreach($params as $v) {
 					$key = array_shift($keys);
 					$this->build[$key] = $v;
 				}
 			}
 			
-			$ret = parent::execute($input_parameters);
+			parent::execute($params);
+
 			if(go()->getDebugger()->enabled && go()->getDbConnection()->debug && isset($this->build)) {
 				$duration  = number_format((go()->getDebugger()->getMicrotime() * 1000) - ($this->build['start'] * 1000), 2);
 
 				$sql = QueryBuilder::debugBuild($this->build);
 				go()->debug(str_replace(["\n","\t"], "", $sql) . "\n(" . $duration . 'ms)', 5);
 			}
-			if($ret === false) {
-				go()->error("SQL FAILURE: " . $this);
-			}
-			return $ret;
+
+			return true;
 		}
-		catch(Exception $e) {
+		catch(PDOException $e) {
 			go()->error("SQL FAILURE: " . $this);
 			throw $e;
 		}
 	}
-// TODO UPDATE PARAMS IN build
-	// public function bindValue($parameter, $value, $data_type = PDO::PARAM_STR)
-	// {
-	// 	if(isset($this->build)) {
-	// 		$this->build['params'][$parameter] = $value;
-	// 	}
-
-	// 	return parent::bindValue($parameter, $value, $data_type);
-	// }
-
-	// public function bindParam($parameter, &$variable, $data_type = PDO::PARAM_STR, $length = null, $driver_options = null)
-	// {
-	// 	if(isset($this->build)) {
-	// 		$this->build['params'][$parameter] = $variable;
-	// 	}
-
-	// 	return parent::bindParam($parameter, $variable, $data_type, $length, $driver_options);
-	// }
 	
 }
