@@ -959,7 +959,6 @@ class User extends Entity {
 	 * the \go\core\auth\DomainProvider interface.
 	 *
 	 * @return string[]
-	 * @throws \go\core\exception\ConfigurationException
 	 */
 	public static function getAuthenticationDomains() {
 		$classes = go()->getCache()->get("authentication-domains-providers");
@@ -1166,7 +1165,7 @@ class User extends Entity {
 		return Token::delete($query) && RememberMe::delete($query);
 	}
 
-	public function toArray($properties = [])
+	public function toArray(array $properties = null): array
 	{
 		return parent::toArray($properties); // TODO: Change the autogenerated stub
 	}
